requirements.txt
streamlit
moviepy
numpy
pillow
proglog
imageio
imageio-ffmpeg
watchdog